rafactor: extracts random
