+++ src/index.js
@@ -1,5 +1,6 @@
 import Konva from 'konva';
 import Ball from './ball.js';
+import { getRandomColor, getRandomVelocity, getRandomNumberInRange } from './random.js';
 
 const width = window.innerWidth;
 const height = window.innerHeight;
@@ -72,31 +73,4 @@ function bounceOnWalls (ball) {
 	if (bounced) {
 		ball.changeColor(getRandomColor());
 	}
-}
-
-function getRandomColor () {
-	const saturation = getRandomNumberInRange(25, 100);
-	const lightness = getRandomNumberInRange(20, 75);
-	const hue = getRandomNumberInRange(0, 360);
-	return `hsl(${hue} ${saturation}% ${lightness}%)`;
-}
-
-function getRandomVelocity () {
-	let dx = 0;
-	let dy = 0;
-	while (dx === 0 && dy === 0) {
-		dx = getRandomNumberInRange(-7, 7);
-		dy = getRandomNumberInRange(-7, 7);
-	}
-	return { dx, dy };
-}
-
-function getRandomNumberInRange (min, max) {
-	if (min > max) {
-		const tmp = min;
-		min = max;
-		max = tmp;
-	}
-	const diff = max - min;
-	return Math.round(Math.random() * diff) + min;
 }
